pyproject.toml
[tool.poetry]
name = "social-app"
version = "0.1.0"
description = "A social media backend"
authors = ["Atrayan"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
types-pyyaml = "^6.0.12.20240808"
django-channels = "^0.7.0"
pymongo = "^4.8.0"
dnspython = "^2.6.1"
drf-yasg = "^1.21.7"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
types-pyyaml = "^6.0.12.20240808"
colorlog = "^6.8.2"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
